fix: downloading instant client based on os type

+++ dev/oracle/latest/start.sh
@@ -18,15 +18,22 @@ docker cp privileges.sql oraclexedb:/opt/oracle/.
 docker exec -it oraclexedb sqlplus system/password@XEPDB1 @privileges.sql
 
 # Setting up Oracle instant client for oracledb
-if [ ! -d  ~/Downloads/oracle ] 
+if [ ! -d  ~/Downloads/oracle ] && [ $(uname) == 'Linux' ]
 then 
     mkdir ~/Downloads/oracle && 
     wget https://download.oracle.com/otn_software/linux/instantclient/215000/instantclient-basic-linux.x64-21.5.0.0.0dbru.zip --no-check-certificate && 
     unzip instantclient-basic-linux.x64-21.5.0.0.0dbru.zip -d ~/Downloads/oracle/ &&
     rm instantclient-basic-linux.x64-21.5.0.0.0dbru.zip &&
-    mv ~/Downloads/oracle/instantclient_21_5 ~/Downloads/oracle/instantclient &&
-    export LD_LIBRARY_PATH=~/Downloads/oracle/instantclient:$LD_LIBRARY_PATH &&
-    export TNS_ADMIN=~/Downloads/oracle/instantclient/network/admin:$TNS_ADMIN
+    mv ~/Downloads/oracle/instantclient_21_5 ~/Downloads/oracle/instantclient
+
+    echo "Local Oracle instant client has been setup!"
+elif [ ! -d  ~/Downloads/oracle ] && [ $(uname) == 'Darwin' ]
+then
+    mkdir ~/Downloads/oracle && 
+    wget https://download.oracle.com/otn_software/mac/instantclient/198000/instantclient-basic-macos.x64-19.8.0.0.0dbru.zip --no-check-certificate && 
+    unzip instantclient-basic-macos.x64-19.8.0.0.0dbru.zip -d ~/Downloads/oracle/ &&
+    rm instantclient-basic-macos.x64-19.8.0.0.0dbru.zip &&
+    mv ~/Downloads/oracle/instantclient_19_8 ~/Downloads/oracle/instantclient
 
     echo "Local Oracle instant client has been setup!"
 fi
